Rename input and form widgets to PopupSearch and EmbeddedSearch
Follow Svelte PascalCase naming convention for all svelte files
