+++ libs/search-widget/src/SearchBar.svelte
