+++ libs/search-widget/src/index.ts
@@ -1,3 +1,3 @@
-export { default as NucliaWidget } from './widget.svelte';
-export { default as NucliaSearchBar } from './search-bar.svelte';
-export { default as NucliaSearchResults } from './search-results.svelte';
+export { default as NucliaWidget } from './Widget.svelte';
+export { default as NucliaSearchBar } from './SearchBar.svelte';
+export { default as NucliaSearchResults } from './SearchResults.svelte';
