+++ libs/search-widget/src/widgets/EmbeddedSearch.spec.ts
@@ -1,12 +1,12 @@
-import FormWidget from './FormWidget.svelte';
+import EmbeddedSearch from './EmbeddedSearch.svelte';
 import { fireEvent, render } from '@testing-library/svelte';
 import { nucliaStore } from '../core/store';
 import { firstValueFrom } from 'rxjs';
 import type { IResource } from '@nuclia/core';
 
-describe('Input widget', () => {
+describe('Embedded search', () => {
   it('should trigger search', async () => {
-    const { container } = render(FormWidget);
+    const { container } = render(EmbeddedSearch);
     const input = container.querySelector('input');
     expect(input).toBeTruthy();
     if (input) {
@@ -22,7 +22,7 @@ describe('Input widget', () => {
         res1: { title: 'Knowledge is power', summary: 'France is bacon' } as IResource,
       },
     });
-    const { container } = render(FormWidget);
+    const { container } = render(EmbeddedSearch);
     expect(container.querySelector('.results:not(.empty)')).toBeTruthy();
   });
 });
