+++ libs/search-widget/src/widgets/PopupSearch.svelte
