+++ libs/search-widget/src/widgets/search-input/SearchInput.spec.ts
